Fixed memory errors that Hudson asked about
Quite involved to track down, but simple enough to fix. There were a lot
of simplifications to overly complex code.

Also changed the way the player's input was being handled (left a
comment to explain).

Also noted an inefficiency in the collision functions. You should not be
using pow for simple squaring and exponents. Instead, use a custom
function if you have to. The important part of this is that pow has a
lot of overhead to make sure it handles fractional and negative
exponents correctly. You don't need it, so it is just slowing down your
code.